Move CLI into separate folder
