+++ drheader/resources/cli/bulk_compare_schema.json
