+++ drheader/resources/cli/bulk_scan_schema.json
